+++ src/main/java/com/quping/entry/UserRatingMapping.java
@@ -18,9 +18,9 @@ import java.util.Date;
 @NoArgsConstructor
 public class UserRatingMapping {
     @TableId(type = IdType.AUTO)
-    private Integer id;
-    private Integer userId;
-    private Integer ratingId;
+    private Long id;
+    private Long userId;
+    private Long ratingId;
     private Integer score;
 
     /**
